rename : Image 관련 도메인 image 패키지로 이동
